pam_ldaphome: optionally run external program after populating home directory.
This allows for dynamic modifications of the initial directory contents,
depending on the login name of the user.  The name of the external program
is given with the initrc-command configuration statement.  It is invoked
with the single argument, specifying the login name.  The standard input
is closed, standard output is diverted to standard error.  Standard error
can be diverted to a file using the initrc-log statement.

* Makefile.am (EXTRA_DIST): Add examples.
* examples/usergitconfig: New file.
* pam_ldaphome/pam_ldaphome.c (run_prog, run_initrc): New statics.
(pam_sm_authenticate): Call run_initrc prior to calling
import_public_key.